04/26 사이드바하고 content 부분 넓이 css 수정 / MydataController만들고, mydata 페이지 mydata폴더에 분리 / controller들 클래스 이름 앞부분 대문자로 변경
