+++ src/main/java/com/ipa/openapi/controller/IndexController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
-public class indexController {
+public class IndexController {
 
     @GetMapping("")
     public String index() {
@@ -29,34 +29,9 @@ public class indexController {
         return "main page";
     }
 
-    @GetMapping("/providerTable")
-    public String providerTable() {
-        return "mdProviderTable";
-    }
-
-    @GetMapping("/CollectorTable")
-    public String collectorTable() {
-        return "mdCollectorTable";
-    }
-
-    @GetMapping("/agencyTable")
-    public String mdServiceTable() {
-        return "mdAgencyTable";
-    }
-
     @GetMapping("/modal")
     public String modal() {
         return "modalTest";
-    }
-
-    @GetMapping("/statistics-Daily")
-    public String statistics_Daily() {
-        return "statistics-Daily";
-    }
-
-    @GetMapping("/statistics-7Day")
-    public String statistics_7Day() {
-        return "statistics-7Day";
     }
 
     @GetMapping("accountList")
@@ -74,13 +49,4 @@ public class indexController {
         return "authorization";
     }
 
-    @GetMapping("mydataServiceControl")
-    public String mydataServiceControl() {
-        return "mydataServiceControl";
-    }
-
-    @GetMapping("/mydataToken")
-    public String mydataToken() {
-        return "mydataToken";
-    }
 }
