+++ src/main/resources/static/css/style.css
@@ -604,7 +604,7 @@ h6 {
     width: 300px;
     z-index: 996;
     transition: all 0.3s;
-    padding: 20px;
+    /*padding: 20px;*/
     overflow-y: auto;
     scrollbar-width: thin;
     scrollbar-color: #aab7cf transparent;
