+++ src/main/resources/templates/thymeleaf/fragment/headSideBar.html
@@ -111,7 +111,7 @@
                         </a>
                     </li>
                     <li>
-                        <a href="#">
+                        <a href="/collectorTable">
                             <i class="bi bi-circle"></i><span>거래이력(사업자)</span>
                         </a>
                     </li>
