+++ src/main/resources/templates/thymeleaf/layout/layout.html
@@ -17,10 +17,10 @@
 
     <th:block th:replace="fragment/headSideBar :: headerFragment"></th:block>
     <div class="d-flex">
-        <div style="width: 12%">
+        <div>
             <th:block th:replace="fragment/headSideBar :: sideBarFragment"></th:block>
         </div>
-        <div class="mt-4" style="float:left; width: 85%">
+        <div class="mt-4">
             <th:block layout:fragment="content"></th:block>
         </div>
     </div>
