+++ src/main/resources/templates/thymeleaf/mydata/mdCollectorTable.html
