+++ src/main/resources/templates/thymeleaf/mydata/statistics-Daily.html
